requirements.txt
newspaper3k
requests
google-cloud-speech
flask